Fix updateStreamlit prop usage on callbacks. Move sendStreamlitData to new file. Clean variables moved to CanvasToolbar. Yes sorry I mixed stuff in this commit
